backend/requirements.txt
flask
flask-cors
python-dotenv
youtube-transcript-api
google-generativeai
